Put code under `PSR7Session\Storageless` namespace
