+++ test/StoragelessTest/Time/SystemCurrentTimeTest.php
@@ -18,10 +18,10 @@
 
 declare(strict_types=1);
 
-namespace PSR7SessionTest\Time;
+namespace PSR7SessionTest\Storageless\Time;
 
 use DateTimeImmutable;
-use PSR7Session\Time\SystemCurrentTime;
+use PSR7Session\Storageless\Time\SystemCurrentTime;
 
 class SystemCurrentTimeTest extends \PHPUnit_Framework_TestCase
 {
